extracted all http-utils shared by lang and nlu server in comon file
