+++ src/bp/http-utils.ts
@@ -1,8 +1,20 @@
+import { Request } from 'express'
 import _ from 'lodash'
 
-import { UnauthorizedError } from '../core/routers/errors'
+import { UnauthorizedError } from './core/routers/errors'
 
 const debugAuth = DEBUG('api:auth')
+
+export const isAdminToken = (req, adminToken: string) => {
+  if (!adminToken || !adminToken.length) {
+    return true
+  }
+  if (!req.headers.authorization) {
+    return false
+  }
+  const [, token] = req.headers.authorization.split(' ')
+  return token === adminToken
+}
 
 export const authMiddleware = (secureToken: string, secondToken?: string) => (req, _res, next) => {
   if (!secureToken || !secureToken.length) {
@@ -33,6 +45,14 @@ export const authMiddleware = (secureToken: string, secondToken?: string) => (re
   next()
 }
 
+export const disabledReadonlyMiddleware = (readonly: boolean) => (_req, _res, next) => {
+  if (readonly) {
+    return next(new UnauthorizedError('API server is running in read-only mode'))
+  }
+
+  next()
+}
+
 export const handleUnexpectedError = (err, req, res, next) => {
   const statusCode = err.statusCode || 500
   const errorCode = err.errorCode || 'BP_000'
@@ -52,4 +72,8 @@ export const handleErrorLogging = (err, req, res, next) => {
   }
 
   next(err)
+}
+
+export type RequestWithLang = Request & {
+  language?: string
 }
